frontend: helpers: Extract getAppUrl to own module
