+++ frontend/src/lib/k8s/api/v1/constants.ts
@@ -1,6 +1,6 @@
-import helpers from '../../../../helpers';
+import { getAppUrl } from '../../../../helpers/getAppUrl';
 
-export const BASE_HTTP_URL = helpers.getAppUrl();
+export const BASE_HTTP_URL = getAppUrl();
 export const CLUSTERS_PREFIX = 'clusters';
 export const JSON_HEADERS = { Accept: 'application/json', 'Content-Type': 'application/json' };
 export const DEFAULT_TIMEOUT = 2 * 60 * 1000; // ms
