+++ frontend/src/plugin/index.ts
@@ -20,7 +20,7 @@ import * as ReactRouter from 'react-router-dom';
 import * as Recharts from 'recharts';
 import semver from 'semver';
 import * as CommonComponents from '../components/common';
-import helpers from '../helpers';
+import { getAppUrl } from '../helpers/getAppUrl';
 import { isElectron } from '../helpers/isElectron';
 import * as K8s from '../lib/k8s';
 import * as ApiProxy from '../lib/k8s/apiProxy';
@@ -246,19 +246,15 @@ export async function fetchAndExecutePlugins(
   onSettingsChange: (plugins: PluginInfo[]) => void,
   onIncompatible: (plugins: Record<string, PluginInfo>) => void
 ) {
-  const pluginPaths = (await fetch(`${helpers.getAppUrl()}plugins`).then(resp =>
-    resp.json()
-  )) as string[];
+  const pluginPaths = (await fetch(`${getAppUrl()}plugins`).then(resp => resp.json())) as string[];
 
   const sourcesPromise = Promise.all(
-    pluginPaths.map(path =>
-      fetch(`${helpers.getAppUrl()}${path}/main.js`).then(resp => resp.text())
-    )
+    pluginPaths.map(path => fetch(`${getAppUrl()}${path}/main.js`).then(resp => resp.text()))
   );
 
   const packageInfosPromise = await Promise.all<PluginInfo>(
     pluginPaths.map(path =>
-      fetch(`${helpers.getAppUrl()}${path}/package.json`).then(resp => {
+      fetch(`${getAppUrl()}${path}/package.json`).then(resp => {
         if (!resp.ok) {
           if (resp.status !== 404) {
             return Promise.reject(resp);
